Security plugin: change policy to advisory
Change policy from enforcing to advisory when access is requested
from unknown domain or permitted to unknown domain. Also trunacated
lines to 80 characters, fixed grammar in the warning messages, and
removed trailing spaces and tabs.

In future, it might make sense to make this behaviour a command-line
option. Added a FIXME comment about that.

Change-Id: Iab78d79a8a28e019ec8601265a59d259fd46f9d6
Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@linux.intel.com>